+++ src/NetMQ.High/ClientTimeout.cs
@@ -7,9 +7,9 @@ namespace NetMQ.High
 {
     public class ClientTimeout : Client
     {
-        readonly double timeout;
+        readonly int timeout;
 
-        public ClientTimeout(double timeout) => this.timeout = timeout;
+        public ClientTimeout(int timeout) => this.timeout = timeout;
 
         // Answer by Lawrence Johnston at https://stackoverflow.com/questions/4238345/asynchronously-wait-for-taskt-to-complete-with-timeout
         static async Task<TResult> TimeoutAfter<TResult>(Task<TResult> task, TimeSpan timeout)
